requirements.txt
numpy
Pillow
Flask
redis